fixed animation and put enum in interface
