+++ SAKPaaS/src/app/components/user-card/user-card.component.scss
@@ -45,7 +45,7 @@
   }
 
   &.expanded {
-    width: auto;
+    width: 440px;
   }
 }
 
@@ -96,7 +96,9 @@
         margin-left: 12.5px;
         margin-top: 12px;
     }
-    overflow: auto;
+    overflow: hidden;
+    display: flex;
+    white-space: nowrap;
     vertical-align: top;
     height: 110px;
   }
@@ -116,8 +118,8 @@
 
   #wrapper {
     overflow: hidden;
-    display:flex;
-    align-items:center;
+    display: flex;
+    align-items: center;
   }
   #first {
     float: left;
